# app/static/swagger.yaml
swagger: "2.0" # swagger version
info:
  title: Mechanic Shop API
  description: API for managing a mechanic shop, including customers, mechanics, inventory, and service tickets
  version: 1.0.0
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Enter the JWT token using the format: Bearer <token>"

# Endpoints
paths:
  /customers/login:
    post:
      tags:
        - "Customers"
      summary: "Login endpoint for customers (generates a JWT token)"
      description: "This endpoint allows customers to log in by providing their email and password. Upon successful authentication, a JWT token is generated and returned in the response."
      parameters:
        - in: "body"
          name: "body"
          description: "Login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Successful login"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              message: "Login successful"
              status: "success"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJleHAiOjE2MzQyMzQwMDB9.abc123"
  /customers: # endpoint for creating a new customer
    post: # Post request
      tags:
        - "Customers"
      summary: "Endpoint to create/register a new customer."
      description: "This endpoint allows the creation of a new customer in the Mechanic Shop system by providing their name, email, phone number, and password. Emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer details required to create a new customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "johndoe@hotmail.com"
              phone: "123-456-7890"
              password: "123password"

    get: # Get request
      tags:
        - "Customers"
      summary: "Endpoint to retrieve all customers."
      description: "This endpoint retrieves a list of all customers in the Mechanic Shop system."
      responses:
        200:
          description: "List of all customers retrieved successfully"
          schema:
            $ref: "#/definitions/AllCustomers"
          examples:
            application/json:
              - id: 1
                name: "John Doe"
                email: "johndoe@hotmail.com"
                phone: "123-456-7890"
                password: "123password"
              - id: 2
                name: "Jane Smith"
                email: "janesmith@hotmail.com"
                phone: "987-654-3210"
                password: "securepassword"
  /customers/{customer_id}:
    get:
      tags:
        - "Customers"
      summary: "Endpoint to retrieve a specific customer's information."
      description: "This endpoint retrieves the information of a specific customer in the Mechanic Shop system based on their unique customer ID."
      parameters:
        - in: "path"
          name: "customer_id"
          description: "ID of the customer to be retrieved."
          required: true
          type: "integer"
      responses:
        200:
          description: "Customer information retrieved successfully"
          schema:
            $ref: "#/definitions/SingleCustomer"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "johndoe@hotmail.com"
              phone: "123-456-7890"
              password: "123password"
  /customers/{customer_id}/update-account:
    put:
      tags:
        - "Customers"
      summary: "Endpoint to update an existing customer's information."
      description: "This endpoint is to update an existing customer's information in the Mechanic Shop system. This is a token authenticated route, and the user must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          description: "ID of the customer to be updated."
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Customer details required to update an existing customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "johndoe@hotmail.com"
              phone: "123-456-7890"
              password: "123password"
  /customers/{customer_id}/delete-account:
    delete:
      tags:
        - "Customers"
      summary: "Endpoint to delete an existing customer's account."
      description: "This endpoint is to delete an existing customer's account in the Mechanic Shop system. This is a token authenticated route, and the user must be logged in to delete their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          description: "ID of the customer to be deleted."
          required: true
          type: "integer"
      responses:
        200:
          description: "Customer account deleted successfully"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer Id: 1 has been deleted successfully"
  /customers/{customer_id}/my-service-tickets:
    get:
      tags:
        - "Customers"
      summary: "Endpoint to retrieve all service tickets for a specific customer."
      description: "This endpoint retrieves a list of all service tickets associated with a specific customer in the Mechanic Shop system based on their unique customer ID. This is a token authenticated route, and the user must be logged in to view their service tickets."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          description: "ID of the customer whose service tickets are to be retrieved."
          required: true
          type: "integer"
      responses:
        200:
          description: "Customer service tickets retrieved successfully"
          schema:
            $ref: "#/definitions/AllServiceTicketsResponse"
          examples:
            application/json:
              - customer:
                  name: "John Doe"
                  email: "johndoe@hotmail.com"
                customer_id: 1
                id: 1
                inventory: [1, 2]
                mechanics: [1, 2]
                service_date: "2024-06-15"
                service_desc: "Oil change and tire rotation"
                vin: "1HGCM82633A123456"
  /mechanics/login:
    post:
      tags:
        - "Mechanics"
      summary: "Login endpoint for mechanics (generates a JWT token)"
      description: "This endpoint allows mechanics to log in by providing their email and password. Upon successful authentication, a JWT token is generated and returned in the response."
      parameters:
        - in: "body"
          name: "body"
          description: "Login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Successful login"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              message: "Login successful"
              status: "success"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJleHAiOjE2MzQyMzQwMDB9.abc123"
  /mechanics: # endpoint for creating a new mechanic
    post: # Post request
      tags:
        - "Mechanics"
      summary: "Endpoint to create/register a new mechanic."
      description: "This endpoint allows the creation of a new mechanic in the Mechanic Shop system by providing their name, email, phone number, salary, and password. Emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic details required to create a new mechanic."
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/CreateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: "Jane Smith"
              email: "jane.smith@example.com"
              phone: "123-456-7890"
              salary: 50000.00
              password: "securepassword"

    get: # Get request
      tags:
        - "Mechanics"
      summary: "Endpoint to retrieve all mechanics."
      description: "This endpoint retrieves a list of all mechanics in the Mechanic Shop system."
      responses:
        200:
          description: "List of all mechanics retrieved successfully"
          schema:
            $ref: "#/definitions/AllMechanics"
          examples:
            application/json:
              - id: 1
                name: "Jane Smith"
                email: "jane.smith@example.com"
                phone: "123-456-7890"
                salary: 50000.00
                password: "securepassword"
              - id: 2
                name: "John Doe"
                email: "john.doe@example.com"
                phone: "987-654-3210"
                salary: 60000.00
                password: "securepassword"
  /mechanics/{mechanic_id}:
    put:
      tags:
        - "Mechanics"
      summary: "Endpoint to update an existing mechanic's information."
      description: "This endpoint is to update an existing mechanic's information in the Mechanic Shop system."
      parameters:
        - in: "path"
          name: "mechanic_id"
          description: "ID of the mechanic to be updated."
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Mechanic details required to update an existing mechanic."
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: "Jane Smith"
              email: "jane.smith@example.com"
              phone: "123-456-7890"
              salary: 50000.00
              password: "securepassword"

    delete:
      tags:
        - "Mechanics"
      summary: "Endpoint to delete an existing mechanic."
      description: "This endpoint is to delete an existing mechanic in the Mechanic Shop system."
      parameters:
        - in: "path"
          name: "mechanic_id"
          description: "ID of the mechanic to be deleted."
          required: true
          type: "integer"
      responses:
        200:
          description: "Mechanic deleted successfully"
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"
          examples:
            application/json:
              message: "Mechanic Id: 1 has been deleted successfully"
              status: "success"
...
definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
      token:
        type: "string"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        password:
          type: "string"
  SingleCustomer:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"
  ServiceTicketCustomer:
    type: "object"
    properties:
      email:
        type: "string"
      name:
        type: "string"
  AllServiceTicketsResponse:
    type: "array"
    items:
      $ref: "#/definitions/ServiceTicketResponse"
  CreateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - salary
      - password
  CreateMechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
  AllMechanics:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        salary:
          type: "number"
          format: "float"
        password:
          type: "string"
  UpdateMechanicPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
  UpdateMechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
        format: "float"
      password:
        type: "string"
  DeleteMechanicResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
...
  ServiceTicketResponse:
    type: "object"
    properties:
      customer:
        $ref: "#/definitions/ServiceTicketCustomer"
      customer_id:
        type: "integer"
      id:
        type: "integer"
      inventory:
        type: "array"
        items:
          type: "integer"
      mechanics:
        type: "array"
        items:
          type: "integer"
      service_date:
        type: "string"
        format: "date"
      service_desc:
        type: "string"
      vin:
        type: "string"